+++ src/test/java/io/github/thijzert123/homewizard4j/test/v1/EnergySocketTest.java
@@ -15,7 +15,7 @@ import java.io.IOException;
  * @author Thijzert123
  */
 public class EnergySocketTest {
-    private static EnergySocket energySocket;
+    private static String expectedJson;
 
     @BeforeAll
     public static void beforeAll() throws IOException {
@@ -23,13 +23,22 @@ public class EnergySocketTest {
         httpServer.createContext("/test/state", new TestHttpHandler(Utils.getResourceAsString("energySocket/state.json")));
         httpServer.start();
 
-        energySocket = new EnergySocket(true, "localhost", 8323, "/test");
-        energySocket.updateAll();
+        expectedJson = "{\"service_name\":null,\"api_enabled\":true,\"host_address\":\"localhost\",\"port\":8323,\"api_path\":\"/test\",\"system_configuration\":{\"cloud_enabled\":true},\"energy_socket_state\":{\"power_on\":true,\"switch_lock\":false,\"brightness\":255},\"product_type\":\"HWE-SKT\",\"product_name\":\"Energy Socket\",\"serial\":\"3c35e7aabbcc\",\"firmware_version\":\"5.18\",\"api_version\":\"v1\",\"wifi_ssid\":\"My Wi-Fi\",\"wifi_strength\":100.0,\"total_power_import_kwh\":30.511,\"total_power_export_kwh\":85.951,\"active_power_w\":543.312,\"active_voltage_v\":231.539,\"active_current_a\":2.346,\"active_reactive_power_var\":123.456,\"active_apparent_power_va\":666.768,\"active_power_factor\":0.81688,\"active_frequency_hz\":50.005}";
     }
 
     @Test
-    public void test() throws HomeWizardApiException {
-        Assertions.assertEquals("{\"service_name\":null,\"api_enabled\":true,\"host_address\":\"localhost\",\"port\":8323,\"api_path\":\"/test\",\"system_configuration\":{\"cloud_enabled\":true},\"energy_socket_state\":{\"power_on\":true,\"switch_lock\":false,\"brightness\":255},\"product_type\":\"HWE-SKT\",\"product_name\":\"Energy Socket\",\"serial\":\"3c35e7aabbcc\",\"firmware_version\":\"5.18\",\"api_version\":\"v1\",\"wifi_ssid\":\"My Wi-Fi\",\"wifi_strength\":100.0,\"total_power_import_kwh\":30.511,\"total_power_export_kwh\":85.951,\"active_power_w\":543.312,\"active_voltage_v\":231.539,\"active_current_a\":2.346,\"active_reactive_power_var\":123.456,\"active_apparent_power_va\":666.768,\"active_power_factor\":0.81688,\"active_frequency_hz\":50.005}",
-                energySocket.toJson());
+    public void testToJson() throws HomeWizardApiException {
+        final EnergySocket energySocket = new EnergySocket(true, "localhost", 8323, "/test");
+        energySocket.updateAll();
+
+        Assertions.assertEquals(expectedJson, energySocket.toJson());
+    }
+
+    @Test
+    public void testFromJson() throws HomeWizardApiException {
+        final EnergySocket energySocket = new EnergySocket(true, "localhost", 8323, "/test");
+        energySocket.updateFromJson(expectedJson);
+
+        Assertions.assertEquals(expectedJson, energySocket.toJson());
     }
 }